pyproject.toml
[project]
name = "nlp-financial-event-clustering-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gensim>=4.4.0",
    "lxml-html-clean>=0.4.4",
    "lxml[html-clean]>=6.0.2",
    "nbformat>=5.10.4",
    "newspaper3k>=0.2.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "peft>=0.18.1",
    "plotly>=6.6.0",
    "requests>=2.33.1",
    "ruff>=0.15.9",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "spacy>=3.8.14",
    "telethon>=1.42.0",
    "tokenizers>=0.22.2",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
    "transformers>=5.5.0",
    "twscrape>=0.17.0",
    "vadersentiment>=3.3.2",
    "yfinance>=1.2.0",
]
